remove airlock_resource_helpers - move the methods to the airlock service
